+++ scripts/generate-fleet-status.py
@@ -6,7 +6,7 @@ import sqlalchemy
 import esdglider.config as config
 
 """
-Make a yaml file for all glider deployments to display on the website
+Write various database views and summaries to the ESD Fleet Status page
 
 Scrape info from database, and generate a yaml file with deployment info.
 Then loop through the deployments, and look at GCP to see what output
@@ -19,9 +19,6 @@ from a local computer to access the database
 'db/glider-db-prod.txt' is the database URL, used to create the
 sqlalchemy engine. It should not be committed to GitHub.
 """
-
-# yaml_path = "C:/Users/sam.woodman/Downloads"
-yaml_path = "../glider-lab-manual/content"
 
 if __name__ == "__main__":
     logging.basicConfig(
@@ -36,6 +33,23 @@ if __name__ == "__main__":
         with engine.connect() as connection:
             # Generate deployment table
             df_depl = config.get_deployment_table(connection)
+            x = df_depl.copy(deep=True)
+            x = x.drop(["Dates", "Sensors"], axis=1)
 
-            # Make website yaml
-            config.make_website_yaml(df_depl, yaml_path)
+            # Write Deployments table to fleet status
+            wk_name = "Deployments-Database"
+            logging.info("Updating the Fleet Status %s sheet", wk_name)
+            x = x.fillna("").rename({"Glider_Deployment_ID": "Deployment_ID"})
+            gc = gspread.oauth()  # type: ignore
+            sh = gc.open("Fleet Status")
+            wk = sh.worksheet(wk_name)
+            wk.update([x.columns.values.tolist()] + x.values.tolist())
+
+            # # Update data validation formatting automatically..
+            # from gspread.utils import ValidationConditionType
+            # wk.add_validation(
+            #     f'F2:L{1+x.shape[0]}',
+            #     ValidationConditionType.one_of_list,
+            #     ['TRUE', 'FALSE'],
+            #     showCustomUi=True
+            # )